refactor: decompose onboarding selections
